[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.01 ms, Memory: 16.3 MB -BaekjoonHub

+++ 프로그래머스/lv0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-func solution(_ balls:Int, _ share:Int) -> Double {
+func solution(_ balls:Int, _ share:Int) -> Int {
     var balls: Int = balls
     var share: Int = share
     var answer: Double = 1
@@ -13,5 +13,5 @@ func solution(_ balls:Int, _ share:Int) -> Double {
         answer /= Double(balls - share - i)
     }
     
-    return answer
+    return Int(answer)
 }
